+++ test/components/Tab_spec.js
@@ -15,6 +15,17 @@ describe('Test of Tab', () => {
     component = TestUtils.renderIntoDocument(<Tab><p>test tab</p></Tab>);
     expect(component.props.title).to.be.equal('untitled');
     expect(component.props.disableClose).to.be.equal(false);
+
+    expect(component.props.tabClassNames).to.be.an('object');
+    expect(component.props.tabClassNames.tab).to.be.equal('');
+    expect(component.props.tabClassNames.tabBefore).to.be.equal('');
+    expect(component.props.tabClassNames.tabAfter).to.be.equal('');
+    expect(component.props.tabClassNames.tabTitle).to.be.equal('');
+    expect(component.props.tabClassNames.tabCloseIcon).to.be.equal('');
+    expect(component.props.tabClassNames.tabActive).to.be.equal('');
+
+    expect(component.props.tabStyles).to.be.empty;
+
   });
 
   it('Tab works like a scala case class, it just render it\'s children', () => {