+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
   * #1 [Improve tab title color](https://github.com/georgeOsdDev/react-draggable-tab/issues/1)
   * #2 [onTabSelected, onTabClosed, onTabPositionChanged should be called in setState callback](https://github.com/georgeOsdDev/react-draggable-tab/issues/2)
+  * #2 [Enable to set style and class on each Tab](https://github.com/georgeOsdDev/react-draggable-tab/issues/3)
 
 ### Ver 0.1.0 Initial release